+++ FirebaseAuth/Sources/Public/FirebaseAuth/FIRGameCenterAuthProvider.h
@@ -41,8 +41,10 @@ typedef void (^FIRGameCenterCredentialCallback)(FIRAuthCredential *_Nullable cre
     NS_SWIFT_NAME(GameCenterCredentialCallback);
 
 /** @class FIRGameCenterAuthProvider
-    @brief A concrete implementation of @c FIRAuthProvider for Game Center Sign In.
+    @brief A concrete implementation of @c FIRAuthProvider for Game Center Sign In. Not available on
+           watchOS.
  */
+API_UNAVAILABLE(watchos)
 NS_SWIFT_NAME(GameCenterAuthProvider)
 @interface FIRGameCenterAuthProvider : NSObject