+++ GoogleDataTransport/GDTCORLibrary/GDTCORFlatFileStorage.m
@@ -58,6 +58,7 @@ NSString *const kGDTCORBatchComponentsExpirationKey = @"GDTCORBatchComponentsExp
   [[GDTCORRegistrar sharedInstance] registerStorage:[self sharedInstance] target:kGDTCORTargetCCT];
   [[GDTCORRegistrar sharedInstance] registerStorage:[self sharedInstance] target:kGDTCORTargetFLL];
   [[GDTCORRegistrar sharedInstance] registerStorage:[self sharedInstance] target:kGDTCORTargetCSH];
+  [[GDTCORRegistrar sharedInstance] registerStorage:[self sharedInstance] target:kGDTCORTargetINT];
 
   // Sets a global translation mapping to decode GDTCORStoredEvent objects encoded as instances of
   // GDTCOREvent instead. Then we do the same thing with GDTCORStorage. This must be done in load