+++ Package.swift
@@ -67,6 +67,10 @@ let package = Package(
       targets: ["FirebaseFunctions"]
     ),
     .library(
+      name: "FirebaseInAppMessaging-Beta",
+      targets: ["FirebaseInAppMessaging"]
+    ),
+    .library(
       name: "FirebaseInstallations",
       targets: ["FirebaseInstallations"]
     ),
@@ -440,6 +444,32 @@ let package = Package(
         .define("FIRFunctions_VERSION", to: "0.0.1"), // TODO: Fix version
       ]
     ),
+
+    .target(
+      name: "FirebaseInAppMessaging",
+      dependencies: [
+        "FirebaseCore",
+        "FirebaseInstallations",
+        "FirebaseABTesting",
+        "GoogleUtilities_Environment",
+        .product(name: "nanopb", package: "nanopb"),
+      ],
+      path: "FirebaseInAppMessaging/Sources",
+      exclude: [
+        "DefaultUI/CHANGELOG.md",
+        "DefaultUI/README.md",
+      ],
+      resources: [.process("Resources")],
+      publicHeadersPath: "Public",
+      cSettings: [
+        .headerSearchPath("../../"),
+        .define("FIRInAppMessaging_LIB_VERSION", to: "0.0.1"), // TODO: Fix version
+        .define("PB_FIELD_32BIT", to: "1"),
+        .define("PB_NO_PACKED_STRUCTS", to: "1"),
+        .define("PB_ENABLE_MALLOC", to: "1"),
+      ]
+    ),
+
     // .target(
     //   name: "FirebaseInstanceID",
     //   dependencies: ["FirebaseCore", "FirebaseInstallations",
@@ -568,7 +598,6 @@ let package = Package(
       dependencies: [
         "FirebaseAuth",
         "FirebaseABTesting",
-        "FirebaseFunctions",
         "Firebase",
         "FirebaseCrashlytics",
         "FirebaseCore",
@@ -576,6 +605,8 @@ let package = Package(
         "FirebaseDynamicLinks",
         "FirebaseFirestore",
         "FirebaseFirestoreSwift",
+        "FirebaseFunctions",
+        "FirebaseInAppMessaging",
         "FirebaseInstallations",
         // "FirebaseInstanceID",
         "FirebaseRemoteConfig",
@@ -600,13 +631,14 @@ let package = Package(
       dependencies: [
         "FirebaseAuth",
         "FirebaseABTesting",
-        "FirebaseFunctions",
         "Firebase",
         "FirebaseCrashlytics",
         "FirebaseCore",
         "FirebaseDatabase",
         "FirebaseDynamicLinks",
         "FirebaseFirestore",
+        "FirebaseFunctions",
+        "FirebaseInAppMessaging",
         "FirebaseInstallations",
         "FirebaseRemoteConfig",
         "FirebaseStorage",