+++ FirebaseAuth/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# v6.9.0
 - [added] Added support for multi-tenancy (#6142).
 - [added] Added basic watchOS support. (#4621)